Treate empty configuration lists as such
Make sure not to turn empty configuration list options into a list with
a single empty string in it.

This prevents a problem with the default value for exclude_folders
containing an empty string. This empty string matches the root directory
(which happens not to have a name, i.e. ""), thus preventing all files
from being matched.